fix: moves user creation to the auth controller
