+++ src/app/routes/Auth/Auth.route.ts
@@ -1,12 +1,10 @@
 import { Router } from 'express';
-
 import AuthController from 'controllers/Auth/AuthController';
-import UserController from 'controllers/User/UserController';
 
 const router = Router();
 
 router.post('/login', AuthController.loginUser);
 router.post('/logout', AuthController.logoutUser);
-router.post('/register', UserController.createUser);
+router.post('/register', AuthController.registerUser);
 
 export default router;
